make sure we cant change playground or id of element
make sure we cant change playground or id of element

+++ main/java/playground/logic/jpa/jpaElementService.java
@@ -258,15 +258,19 @@ public class jpaElementService implements ElementService {
 	@MyLog
 	@ManagerLogin
 	public void replaceElementWith(String userPlayground, String email, ElementEntity entity, String id, String creatorplayground) {
-		ElementEntity tempElement = this.getElement(userPlayground, email, ElementEntity.createKey(id, creatorplayground));
-		System.err.println("tempElement: " + tempElement);
-		if (tempElement != null) {
-			// Deletes old and replaces with new
-			entity.setCreationDate(tempElement.getCreationDate());
-			elementsDB.deleteById(tempElement.getSuperkey());
-			elementsDB.save(entity);
-		} else
-			throw new ElementDataException("element data for update is incorrect");
+		if(!entity.getId().equals(id) || !entity.getCreatorPlayground().equals(creatorplayground))
+			throw new ElementDataException("Cannot change users Id or creatorplayground");
+		else {
+			ElementEntity tempElement = this.getElement(userPlayground, email, ElementEntity.createKey(id, creatorplayground));
+			System.err.println("tempElement: " + tempElement);
+			if (tempElement != null) {
+				// Deletes old and replaces with new
+				entity.setCreationDate(tempElement.getCreationDate());
+				elementsDB.deleteById(tempElement.getSuperkey());
+				elementsDB.save(entity);
+			} else
+				throw new ElementDataException("element data for update is incorrect");
+		}
 	}
 
 	@Override
